Add sampling rate setting to evaluation triggers
Adds a configurable sampling rate to evaluation trigger settings, allowing users to reduce the percentage of eligible responses that are evaluated. Previously, all matching responses were evaluated unconditionally. Now users can set a rate between 1-100% (defaulting to 100%) to control evaluation frequency, useful for reducing cost and load on high-traffic prompts.

Sampling is applied during live evaluation only — batch experiment evaluations remain deterministic and always evaluate all selected spans. The sampling check runs after trigger target filtering (first/every/last) so it composes naturally with existing trigger logic.
