+++ packages/constants/src/evaluations/shared.ts
@@ -30,14 +30,23 @@ export const DEFAULT_LAST_INTERACTION_DEBOUNCE_SECONDS = 120
 export const LAST_INTERACTION_DEBOUNCE_MIN_SECONDS = 30
 export const LAST_INTERACTION_DEBOUNCE_MAX_SECONDS = 60 * 60 * 24 // 1 day
 
+export const DEFAULT_EVALUATION_SAMPLE_RATE = 100
+export const MIN_EVALUATION_SAMPLE_RATE = 1
+export const MAX_EVALUATION_SAMPLE_RATE = 100
+
 const triggerConfiguration = z.object({
   target: z.enum(EVALUATION_TRIGGER_TARGETS),
   lastInteractionDebounce: z
     .number()
     .min(LAST_INTERACTION_DEBOUNCE_MIN_SECONDS)
     .max(LAST_INTERACTION_DEBOUNCE_MAX_SECONDS)
-    .optional()
-    .default(DEFAULT_LAST_INTERACTION_DEBOUNCE_SECONDS),
+    .optional(),
+  sampleRate: z
+    .number()
+    .int()
+    .min(MIN_EVALUATION_SAMPLE_RATE)
+    .max(MAX_EVALUATION_SAMPLE_RATE)
+    .optional(),
 })
 export type TriggerConfiguration = z.infer<typeof triggerConfiguration>
