+++ packages/core/drizzle/0269_evaluation_trigger_sample_rate.sql
@@ -1,0 +1,5 @@
+UPDATE "latitude"."evaluation_versions"
+SET configuration = jsonb_set("configuration", '{trigger,sampleRate}', '100'::jsonb)
+WHERE configuration IS NOT NULL
+  AND configuration->'trigger' IS NOT NULL
+  AND (configuration->'trigger'->>'sampleRate') IS NULL;
